- patterns: migrated to x64

+++ tools/patterns/jpegload.h
@@ -3,9 +3,9 @@ int JpegLoad(
     char *filename,
     void(*AddScanlineData) (unsigned char *buffer, int stride, void *Param),
     unsigned char **DecodeBuffer,
-    int *DecodeBufferSize,
-    int *DecodeWidth,
-    int *DecodeHegith,
+    long *DecodeBufferSize,
+    long *DecodeWidth,
+    long *DecodeHegith,
     void *Param);
 
 HBITMAP CreateBitmapFromPixels(HDC hDC, UINT uWidth, UINT uHeight, UINT uBitsPerPixel, PBYTE pBits);
